api/requirements.txt
# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
clickhouse-driver==0.2.7
python-dateutil==2.8.2
